src/main/java/com/uallace/ecommerce/models/Pedido.java
package com.uallace.ecommerce.models;


public class Pedido {
    
    public int idCliente;
    public Double valorTotal;
    public int getIdCliente() {
        return idCliente;
    }
    public void setIdCliente(int idCliente) {
        this.idCliente = idCliente;
    }
    public Double getValorTotal() {
        return valorTotal;
    }
    public void setValorTotal(Double valorTotal) {
        this.valorTotal = valorTotal;
    }
    public String getData() {
        return data;
    }
    public void setData(String data) {
        this.data = data;
    }
    public String data;
}
